requirements.txt
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
tqdm>=4.65.0
scipy>=1.10.0
nltk>=3.8.1
sentencepiece>=0.1.99
anthropic>=0.8.1  # Claude API client
tenacity>=8.2.3   # For retry logic
python-dotenv>=1.0.0  # For environment variables
